pyproject.toml
[project]
name = "technical-assessment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.35.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
